wip: cobra cli refactor -- section cmd + docs
Remove unused cobra command get/section and use flag instead.
Removed as the functionality achieved by this command is better
placed as a persistent flag in the get command (flag cascades to
child commands get/page and future get/pages). Will allow users
to specify sections to parse on multiple pages instead of having
to build a unique command or flag on the get/pages to achieve that.
Documentation and some var names for the changed files have also
been updated to improve consistency and developer experience.