wip: cobra cli refactor -- add pages cmd
Add the pages command as a subcommand of get. Allows for users to
scrape a list of pages whose names are defined in a file. The persistent
section flag allows for the user to also scrape only specific sections
from all pages. At the moment only json input is supported.